Fix stack buffer overflow.
This fixes a stack buffer overflow in packing.hpp. The problem is that
we divide by four and round up, and then we multiply by four. So, we
pretend that an array of length 19 has length 20. This probably does not
cause a segfault because it is only a small overflow, and maybe the data
that gets corrupted is not even needed anymore. However, this is still
UB.

This is the Asan report:
=================================================================
==175335==ERROR: AddressSanitizer: stack-buffer-overflow on address 0x7fffc62043e3 at pc 0x00000040952f bp 0x7fffc62042e0 sp 0x7fffc62042d8
WRITE of size 1 at 0x7fffc62043e3 thread T0
    #0 0x40952e in unpackKmer(unsigned char const*, char*) /global/homes/p/petervan/hw3-fork/packing.hpp:104
    #1 0x409654 in pkmer_t::get[abi:cxx11]() const /global/homes/p/petervan/hw3-fork/pkmer_t.hpp:27
    #2 0x4099e1 in kmer_pair::kmer_str[abi:cxx11]() const /global/homes/p/petervan/hw3-fork/kmer_t.hpp:47
    #3 0x409c0e in kmer_pair::next_kmer() const /global/homes/p/petervan/hw3-fork/kmer_t.hpp:52
    #4 0x40dcae in main /global/homes/p/petervan/hw3-fork/kmer_hash.cpp:102
    #5 0x151e93be729c in __libc_start_main (/lib64/libc.so.6+0x3529c)
    #6 0x408c79 in _start (/global/u2/p/petervan/hw3-fork/build/kmer_hash_19+0x408c79)

Address 0x7fffc62043e3 is located in stack of thread T0 at offset 83 in frame
    #0 0x4095b1 in pkmer_t::get[abi:cxx11]() const /global/homes/p/petervan/hw3-fork/pkmer_t.hpp:25

  This frame has 2 object(s):
    [48, 49) '<unknown>'
    [64, 83) 'kmer' (line 26) <== Memory access at offset 83 overflows this variable